chore: introduce @trezor/metadata package
