+++ packages/metadata/tsconfig.json
@@ -2,6 +2,8 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
-        { "path": "../../packages/metadata" }
+        { "path": "../env-utils" },
+        { "path": "../utils" },
+        { "path": "../suite-desktop-api" }
     ]
 }
