+++ packages/suite/package.json
@@ -84,7 +84,6 @@
         "@trezor/utils": "workspace:*",
         "bs58check": "^4.0.0",
         "date-fns": "^2.30.0",
-        "dropbox": "^10.34.0",
         "file-saver": "^2.0.5",
         "framer-motion": "^11.0.3",
         "history": "^4.10.1",
